requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
# psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-decouple==3.8
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
user-agents==2.2.0
